Refactor TeamView into modular subcomponents
